+++ demos/manual/src/test/java/is/codion/framework/demos/manual/store/domain/StoreTest.java
@@ -63,7 +63,7 @@ public class StoreTest extends DomainTest {
 		}
 
 		@Override
-		public Optional<Entity> foreignKey(ForeignKey foreignKey) {
+		public Optional<Entity> entity(ForeignKey foreignKey) {
 			//see if the currently running test requires an ADDRESS entity
 			if (foreignKey.referencedType().equals(Address.TYPE)) {
 				return Optional.of(entities().builder(Address.TYPE)
@@ -73,7 +73,7 @@ public class StoreTest extends DomainTest {
 								.build());
 			}
 
-			return super.foreignKey(foreignKey);
+			return super.entity(foreignKey);
 		}
 
 		@Override